requirements.txt
matplotlib>=3.1.1
numpy>=1.16.4
Pillow>=6.1.0
pywin32>=227
